move Checker to utils too
