+++ lib/utils.mjs
@@ -1,5 +1,3 @@
-import { Checker } from './checker'
-
 /**
  * Loops a function n times passing the current loop number to it
  * if the function returns something (other that undefined) it breaks
@@ -16,6 +14,42 @@ export function times(n, fn) {
 
 export const times9 = fn => times(9, fn)
 export const times81 = fn => times(81, fn)
+
+class Checker {
+  constructor() {
+    this.vals = new Array(10)
+    this.reset()
+  }
+
+  reset() {
+    this.vals.fill(true, 1)
+    this.remaining = 9
+  }
+
+  getRemaining() {
+    const { vals } = this
+    const ret = []
+    for (let i = 1; i <= 9; i++ ) {
+      if (vals[i]) {
+        ret.push(i)
+      }
+    }
+    return ret
+    // TODO: return this.vals.filter(Boolean)
+  }
+
+  eliminate(n) {
+    const { vals } = this
+    if (vals[n]) {
+      vals[n] = false
+      this.remaining--
+    }
+    // returns true if there is nothing remaining
+    if (this.remaining === 0) {
+      return true
+    }
+  }
+}
 
 export function setCharAt(str, index, newChar) {
   const before = str.substring(0, index)
